Major refactoring into new packages
