+++ README.md
@@ -6,8 +6,12 @@
 
 # gorex
 
-The `gorex` library provides a client implementation for the [rexOS](https://www.rexos.org) API in Go. The library can
-easily be integrated into your Go project. It can help you to get started with the provided REX API.
+The `gorex` library provides a library which works with [rexOS](https://www.rexos.org). The library can
+easily be integrated into your Go project. It can help you to get started with REX as a developer. The library offers
+two different main features:
+
+* Working with the [REX file format](https://github.com/roboticeyes/openrex/blob/master/doc/rex-spec-v1.md)
+* Working with the [rexOS REST API](https://support.robotic-eyes.com/rest/index.html)
 
 ## Installation
 
@@ -102,4 +106,4 @@ fmt.Println(project)
 
 * [rexOS](https://www.rexos.org)
 * [REX](https://rex.robotic-eyes.com)
-* [REXv1 file format](https://github.com/roboticeyes/openrex/blob/master/doc/rex-spec-v1.md)
+* [REX file format v1](https://github.com/roboticeyes/openrex/blob/master/doc/rex-spec-v1.md)
