+++ http/rexos/gorex.go
@@ -1,12 +1,12 @@
 // Copyright 2019 Robotic Eyes. All rights reserved.
 
-// Package gorex provides a library for accessing the REX API. REX is a cloud-based operating system
+// Package rexos provides a library for accessing the REX API. REX is a cloud-based operating system
 // for building augmented reality applications.
 //
 // The first thing you have to do is to register at https://rex.robotic-eyes.com for a free REX
 // account.  Once you activated your account, you can simply create an API access token with a
 // `ClientId` and a `ClientSecret`.
-package gorex
+package rexos
 
 import (
 	"strings"
