+++ http/rexos/reference.go
@@ -1,6 +1,6 @@
 // Copyright 2019 Robotic Eyes. All rights reserved.
 
-package gorex
+package rexos
 
 // Reference is a spatial anchor which can be attached to a project or a project file.
 type Reference struct {
